requirements/base.txt
django==5.1.6
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
pillow==11.1.0
psycopg2==2.9.10
